refactor: monalisten into a single Cog
This required commits logic to be split in two.
