--- app/components/github_integration/__init__.py
@@ -1,1 +1,0 @@
-from .entities import fmt as fmt
