--- app/components/github_integration/comments/__init__.py
@@ -1,3 +1,0 @@
-from .integration import GitHubComments
-
-__all__ = ("GitHubComments",)
