--- app/components/github_integration/entities/__init__.py
@@ -1,3 +1,0 @@
-from .integration import GitHubEntities
-
-__all__ = ("GitHubEntities",)
